# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Neuralis AI — Multi-feature AI platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /wikipedia/search:
    get:
      operationId: searchWikipedia
      tags: [wikipedia]
      summary: Search Wikipedia for articles
      parameters:
        - name: query
          in: query
          required: true
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WikipediaSearchResult"
  /wikipedia/summary:
    get:
      operationId: getWikipediaSummary
      tags: [wikipedia]
      summary: Get Wikipedia article summary with AI enhancement
      parameters:
        - name: title
          in: query
          required: true
          schema:
            type: string
          description: Wikipedia article title
      responses:
        "200":
          description: Article summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WikipediaSummary"
        "404":
          description: Article not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"

  # ──────────────── Healthcare ────────────────
...
  /study/sessions:
    get:
      operationId: listStudySessions
      tags: [study]
      summary: List all study sessions
      responses:
        "200":
          description: List of study sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StudySession"
    post:
      operationId: createStudySession
      tags: [study]
      summary: Create a new study session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStudySessionBody"
      responses:
        "201":
          description: Created session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudySession"
  /study/sessions/{id}:
    get:
      operationId: getStudySession
      tags: [study]
      summary: Get a study session with all content
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Study session with content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudySessionWithContent"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"
    delete:
      operationId: deleteStudySession
      tags: [study]
      summary: Delete a study session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /study/sessions/{id}/flashcards:
    get:
      operationId: listFlashcards
      tags: [study]
      summary: List flashcards for a session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Flashcards
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Flashcard"
    post:
      operationId: generateFlashcards
      tags: [study]
      summary: Generate AI flashcards from text
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateFlashcardsBody"
      responses:
        "201":
          description: Generated flashcards
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Flashcard"
  /study/sessions/{id}/quiz:
    post:
      operationId: generateQuiz
      tags: [study]
      summary: Generate a quiz from study material
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateQuizBody"
      responses:
        "201":
          description: Generated quiz questions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QuizQuestion"
  /study/summarize:
    post:
      operationId: summarizeText
      tags: [study]
      summary: Summarize and explain a text or topic
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SummarizeTextBody"
      responses:
        "200":
          description: Summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SummarizeTextResponse"
  /study/explain:
    post:
      operationId: explainConcept
      tags: [study]
      summary: Get a detailed AI explanation of a concept
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExplainConceptBody"
      responses:
        "200":
          description: SSE stream of explanation
          content:
            text/event-stream: {}
...
components:
  schemas:
...
    OpenaiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # ── Weather Schemas ──
...
    WikipediaSearchResult:
      type: object
      properties:
        title:
          type: string
        excerpt:
          type: string
        url:
          type: string
        thumbnail:
          type: ["string", "null"]
      required:
        - title
        - excerpt
        - url
    WikipediaSummary:
      type: object
      properties:
        title:
          type: string
        extract:
          type: string
        aiSummary:
          type: string
        url:
          type: string
        thumbnail:
          type: ["string", "null"]
        relatedTopics:
          type: array
          items:
            type: string
      required:
        - title
        - extract
        - aiSummary
        - url
        - relatedTopics

    # ── Healthcare Schemas ──
...
    StudySession:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        subject:
          type: string
        flashcardCount:
          type: integer
        quizCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - subject
        - flashcardCount
        - quizCount
        - createdAt
        - updatedAt
    StudySessionWithContent:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        subject:
          type: string
        flashcardCount:
          type: integer
        quizCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        flashcards:
          type: array
          items:
            $ref: "#/components/schemas/Flashcard"
        quizQuestions:
          type: array
          items:
            $ref: "#/components/schemas/QuizQuestion"
      required:
        - id
        - title
        - subject
        - flashcardCount
        - quizCount
        - createdAt
        - updatedAt
        - flashcards
        - quizQuestions
    CreateStudySessionBody:
      type: object
      properties:
        title:
          type: string
        subject:
          type: string
      required:
        - title
        - subject
    Flashcard:
      type: object
      properties:
        id:
          type: integer
        sessionId:
          type: integer
        front:
          type: string
        back:
          type: string
        difficulty:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - sessionId
        - front
        - back
        - difficulty
        - createdAt
    QuizQuestion:
      type: object
      properties:
        id:
          type: integer
        sessionId:
          type: integer
        question:
          type: string
        options:
          type: array
          items:
            type: string
        correctIndex:
          type: integer
        explanation:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - sessionId
        - question
        - options
        - correctIndex
        - explanation
        - createdAt
    GenerateFlashcardsBody:
      type: object
      properties:
        text:
          type: string
        count:
          type: integer
      required:
        - text
    GenerateQuizBody:
      type: object
      properties:
        text:
          type: string
        count:
          type: integer
        difficulty:
          type: string
      required:
        - text
    SummarizeTextBody:
      type: object
      properties:
        text:
          type: string
        style:
          type: string
          enum: [brief, detailed, bullet, academic]
      required:
        - text
    SummarizeTextResponse:
      type: object
      properties:
        summary:
          type: string
        keyPoints:
          type: array
          items:
            type: string
        wordCount:
          type: integer
      required:
        - summary
        - keyPoints
        - wordCount
    ExplainConceptBody:
      type: object
      properties:
        concept:
          type: string
        level:
          type: string
          enum: [beginner, intermediate, advanced, expert]
        subject:
          type: ["string", "null"]
      required:
        - concept
        - level
...