# lib/api-zod/src/generated/types/explainConceptBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Neuralis AI — Multi-feature AI platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ExplainConceptBodyLevel } from "./explainConceptBodyLevel";

export interface ExplainConceptBody {
  concept: string;
  level: ExplainConceptBodyLevel;
  /** @nullable */
  subject?: string | null;
}