# lib/api-zod/src/generated/types/generateQuizBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Neuralis AI — Multi-feature AI platform API
 * OpenAPI spec version: 0.1.0
 */

export interface GenerateQuizBody {
  text: string;
  count?: number;
  difficulty?: string;
}