# lib/api-zod/src/generated/types/studySession.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Neuralis AI — Multi-feature AI platform API
 * OpenAPI spec version: 0.1.0
 */

export interface StudySession {
  id: number;
  title: string;
  subject: string;
  flashcardCount: number;
  quizCount: number;
  createdAt: Date;
  updatedAt: Date;
}