# lib/api-zod/src/generated/types/studySessionWithContent.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Neuralis AI — Multi-feature AI platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Flashcard } from "./flashcard";
import type { QuizQuestion } from "./quizQuestion";

export interface StudySessionWithContent {
  id: number;
  title: string;
  subject: string;
  flashcardCount: number;
  quizCount: number;
  createdAt: Date;
  updatedAt: Date;
  flashcards: Flashcard[];
  quizQuestions: QuizQuestion[];
}